feat: Implement initial project structure

+++ WordMaster.Application/ViewModels/Folder/FolderDetailViewModel.cs
@@ -2,7 +2,7 @@ namespace WordMaster.Application.ViewModels.Folder;
 
 public class FolderDetailViewModel
 {
-    public int FolderId { get; set; }
+    public long FolderId { get; set; }
     public string FolderName { get; set; } = string.Empty;
 
     public List<WordMaster.Domain.Entities.Word> Words { get; set; } = new();
