+++ WordMaster.Domain/Entities/WordFolder.cs
@@ -2,10 +2,10 @@ namespace WordMaster.Domain.Entities;
 
 public class WordFolder
 {
-    public int WordId { get; set; }
+    public long WordId { get; set; }
     public Word Word { get; set; } = null!;
 
-    public int FolderId { get; set; }
+    public long FolderId { get; set; }
     public Folder Folder { get; set; } = null!;
 }
